requirements.txt
# COSMOS Framework Dependencies
# Install: pip install -r requirements.txt

# Core dependencies
torch>=1.10.0
numpy>=1.20.0
gymnasium>=0.28.0
scipy>=1.7.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Visualization
matplotlib>=3.5.0
pillow>=9.0.0

# Logging (optional)
wandb>=0.15.0

# Development (optional)
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.0.0
